Solution for part3-5 and part3-6
part3.5 Expand the backend so that new phonebook entries can be added by making HTTP POST requests to the address http://localhost:3001/api/persons.
Generate a new id for the phonebook entry with the Math.random function. 

part3.6 Implement error handling for creating new entries. The request is not allowed to succeed, if:
The name or number is missing
The name already exists in the phonebook